+++ src/main/java/com/verdantartifice/thaumicwonders/client/config/KeyBindings.java
@@ -1,7 +1,11 @@
 package com.verdantartifice.thaumicwonders.client.config;
 
 import com.verdantartifice.thaumicwonders.ThaumicWonders;
+import com.verdantartifice.thaumicwonders.common.entities.EntityFlyingCarpet;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.settings.KeyBinding;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraftforge.client.settings.IKeyConflictContext;
 import net.minecraftforge.fml.client.registry.ClientRegistry;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
@@ -13,11 +17,24 @@ public class KeyBindings {
     public static KeyBinding carpetForwardKey;
     public static KeyBinding carpetBackwardKey;
 
+    private static IKeyConflictContext flyingCarpetContext = new IKeyConflictContext() {
+        @Override
+        public boolean isActive() {
+            EntityPlayer player = Minecraft.getMinecraft().player;
+            return player != null && player.getRidingEntity() instanceof EntityFlyingCarpet;
+        }
+
+        @Override
+        public boolean conflicts(IKeyConflictContext other) {
+            return other.isActive() && this.isActive();
+        }
+    };
+
     public static void init() {
-        carpetForwardKey = new KeyBinding("key.carpet_forward", Keyboard.KEY_W, KEY_CATEGORY);
+        carpetForwardKey = new KeyBinding("key.carpet_forward", flyingCarpetContext, Keyboard.KEY_W, KEY_CATEGORY);
         ClientRegistry.registerKeyBinding(carpetForwardKey);
 
-        carpetBackwardKey = new KeyBinding("key.carpet_backward", Keyboard.KEY_S, KEY_CATEGORY);
+        carpetBackwardKey = new KeyBinding("key.carpet_backward", flyingCarpetContext, Keyboard.KEY_S, KEY_CATEGORY);
         ClientRegistry.registerKeyBinding(carpetBackwardKey);
     }
 }